requirements.txt
streamlit==1.50.0
anthropic==0.96.0
plotly==5.24.1
pandas==2.3.3
requests==2.32.5
reportlab==4.4.10
python-dotenv==1.2.1
supabase==2.18.1
openpyxl==3.1.5
resend==2.16.0
playwright==1.51.0
